Remove deprecated ProjectAnalysisService
