+++ server/sonar-server/src/test/java/org/sonar/server/projectanalysis/ws/DeleteActionTest.java
@@ -40,7 +40,7 @@ import static org.assertj.core.api.Assertions.tuple;
 import static org.sonar.db.component.SnapshotDto.STATUS_PROCESSED;
 import static org.sonar.db.component.SnapshotDto.STATUS_UNPROCESSED;
 import static org.sonar.db.component.SnapshotTesting.newAnalysis;
-import static org.sonarqube.ws.client.projectanalysis.ProjectAnalysesWsParameters.PARAM_ANALYSIS;
+import static org.sonar.server.projectanalysis.ws.ProjectAnalysesWsParameters.PARAM_ANALYSIS;
 
 public class DeleteActionTest {
   @Rule
